DEV: setting up script in utils to add airport data to flights.db
