+++ src/html/glasgow-flight-destinations.html
@@ -63,7 +63,7 @@
   </tr>
   <tr>
     <td><a href="https://www.google.com/maps/place/Alicante">Alicante</a></td>
-    <td><a href="https://www.google.com/search?q=weather+Alicante">6.44</a></td>
+    <td><a href="https://www.google.com/search?q=weather+Alicante">6.29</a></td>
     <td><a href="https://www.skyscanner.de/transport/fluge/gla/ALC/?adults=1&adultsv2=1&cabinclass=economy&children=0&inboundaltsenabled=false&infants=0&outboundaltsenabled=false&preferdirects=true&ref=home&rtn=1">SkyScanner</a></td>
     <td><a href="https://www.airbnb.de/s/Alicante/homes?adults=1&checkin=2024-03-09&checkout=2024-03-12&flexible_trip_lengths%5B%5D=one_week&price_filter_num_nights=3&price_max=112">Airbnb</a> <a href="https://www.booking.com/searchresults.en-gb.html?ss=Alicante&group_adults=1&no_rooms=1&group_children=0&nflt=price%3DEUR-min-110-1%3Breview_score%3D80&flex_window=2&checkin=2024-03-09&checkout=2024-03-12">Booking.com</a></td>
     <td><a href="https://www.google.com/search?q=things+to+do+this+weekend+Alicante">Google Results</a></td>
@@ -143,7 +143,7 @@
   </tr>
   <tr>
     <td><a href="https://www.google.com/maps/place/Madrid">Madrid</a></td>
-    <td><a href="https://www.google.com/search?q=weather+Madrid">4.31</a></td>
+    <td><a href="https://www.google.com/search?q=weather+Madrid">4.17</a></td>
     <td><a href="https://www.skyscanner.de/transport/fluge/gla/MAD/?adults=1&adultsv2=1&cabinclass=economy&children=0&inboundaltsenabled=false&infants=0&outboundaltsenabled=false&preferdirects=true&ref=home&rtn=1">SkyScanner</a></td>
     <td><a href="https://www.airbnb.de/s/Madrid/homes?adults=1&checkin=2024-03-09&checkout=2024-03-12&flexible_trip_lengths%5B%5D=one_week&price_filter_num_nights=3&price_max=112">Airbnb</a> <a href="https://www.booking.com/searchresults.en-gb.html?ss=Madrid&group_adults=1&no_rooms=1&group_children=0&nflt=price%3DEUR-min-110-1%3Breview_score%3D80&flex_window=2&checkin=2024-03-09&checkout=2024-03-12">Booking.com</a></td>
     <td><a href="https://www.google.com/search?q=things+to+do+this+weekend+Madrid">Google Results</a></td>
@@ -351,7 +351,7 @@
   </tr>
   <tr>
     <td><a href="https://www.google.com/maps/place/Lyon">Lyon</a></td>
-    <td><a href="https://www.google.com/search?q=weather+Lyon">3.11</a></td>
+    <td><a href="https://www.google.com/search?q=weather+Lyon">3.09</a></td>
     <td><a href="https://www.skyscanner.de/transport/fluge/gla/LYS/?adults=1&adultsv2=1&cabinclass=economy&children=0&inboundaltsenabled=false&infants=0&outboundaltsenabled=false&preferdirects=true&ref=home&rtn=1">SkyScanner</a></td>
     <td><a href="https://www.airbnb.de/s/Lyon/homes?adults=1&checkin=2024-03-09&checkout=2024-03-12&flexible_trip_lengths%5B%5D=one_week&price_filter_num_nights=3&price_max=112">Airbnb</a> <a href="https://www.booking.com/searchresults.en-gb.html?ss=Lyon&group_adults=1&no_rooms=1&group_children=0&nflt=price%3DEUR-min-110-1%3Breview_score%3D80&flex_window=2&checkin=2024-03-09&checkout=2024-03-12">Booking.com</a></td>
     <td><a href="https://www.google.com/search?q=things+to+do+this+weekend+Lyon">Google Results</a></td>
@@ -399,7 +399,7 @@
   </tr>
   <tr>
     <td><a href="https://www.google.com/maps/place/Stockholm">Stockholm</a></td>
-    <td><a href="https://www.google.com/search?q=weather+Stockholm">2.39</a></td>
+    <td><a href="https://www.google.com/search?q=weather+Stockholm">2.36</a></td>
     <td><a href="https://www.skyscanner.de/transport/fluge/gla/ARN/?adults=1&adultsv2=1&cabinclass=economy&children=0&inboundaltsenabled=false&infants=0&outboundaltsenabled=false&preferdirects=true&ref=home&rtn=1">SkyScanner</a></td>
     <td><a href="https://www.airbnb.de/s/Stockholm/homes?adults=1&checkin=2024-03-09&checkout=2024-03-12&flexible_trip_lengths%5B%5D=one_week&price_filter_num_nights=3&price_max=112">Airbnb</a> <a href="https://www.booking.com/searchresults.en-gb.html?ss=Stockholm&group_adults=1&no_rooms=1&group_children=0&nflt=price%3DEUR-min-110-1%3Breview_score%3D80&flex_window=2&checkin=2024-03-09&checkout=2024-03-12">Booking.com</a></td>
     <td><a href="https://www.google.com/search?q=things+to+do+this+weekend+Stockholm">Google Results</a></td>
